+++ src/components/Common/ProductCard/ProductCard.tsx
@@ -1,12 +1,13 @@
 import { Link } from 'react-router-dom';
-import { LuHeart } from 'react-icons/lu';
 
 import defaultImage from '../../../assets/img/no-image.png';
 import { Product } from '../../../types/interfaces/Product';
-import ButtonAddToCart from '../Buttons/ButtonAddToCart';
+import { Rating } from '../Raiting/Rating';
+import { InStock } from '../InStock/InStock';
+import { FavoriteBtn } from '../Buttons/Favorite/FavoriteBtn';
+import { ButtonAddToCart } from '../Buttons/AddToCart/ButtonAddToCart';
 
 import classes from './ProductCard.module.scss';
-import { Rating } from '../Raiting/Rating';
 
 interface ProductCartProps {
   product: Product;
@@ -22,24 +23,24 @@ export const ProductCard: React.FC<ProductCartProps> = ({ product }) => {
           </div>
         )}
 
-        <div className={classes.img__wrapper}>
+        <div className={classes.card__img_wrapper}>
           {product.images.length > 0 ? (
             <img src={product.images[0].image} alt={product.title} width={254} height={180} />
           ) : (
             <img src={defaultImage} alt="default Image" width={254} height={180} />
           )}
-          {/* <button className={classes.wish}>
-            {product.favorite ? <LuHeart className={classes.wish_img} /> : <LuHeart className={classes.wish_favimg} />}
-          </button> */}
-          <button className={classes.wish}>
-            <LuHeart className={classes.wish_img} />
-          </button>
+          <FavoriteBtn />
         </div>
-        <Rating product={product} />
-        <h4 className={classes.title}>{product.title}</h4>
-        <p className={classes.description}>{product.description}</p>
-        {product.price && <p className={classes.salePrice}>{product.price} грн</p>}
-        <p className={classes.price}>{product.price} грн</p>
+
+        <div className={classes.rating_instock_wrapper}>
+          <Rating product={product} />
+          <InStock product={product} />
+        </div>
+
+        <h4 className={classes.card__title}>{product.title}</h4>
+        <p className={classes.card__description}>{product.description}</p>
+        {product.price && <p className={classes.card__salePrice}>{product.price} грн</p>}
+        <p className={classes.card__price}>{product.price} грн</p>
 
         <ButtonAddToCart />
       </Link>
